TreeView Prototype Implementation
Hardcoded a simple tree view prototype to work on tree view styling
and work through other tree view functionality concepts. This is
largely due to the fact that I'm not entirely sure how I want this
functionality to work, and creating a prototpye seemed to be the most
natural thing to do.

As of now it's simulating a project tree, which "could" be the destination
for drag events for code generation.