Visual Studio Interop Prototyping
Started to scaffold out a simple proxy for the visual studio solution
object exposed by the extensibility model of visual studio... it's a
hot mess to work with natively, so I'm looking into options for making
the vs extensibility objects, specifically the ones that work with
navigating solutions, projects and code easier to work with.

Also created a stub of a method for unit testing this work.