Unit Tests Refactored and Updated
Moved the out of the box unit test library classes to their own
assembly and updated the two unit test projects to have references
to the new class library project. This allows us to leverage this
functionality from numerous test projects.

Also updated one of the test methods to create a solution and it's
projects dynamically rather than rely on an on-disk solution and its
projects. This is a far more flexible approach to unit testing the
extension package.